07fastapi-blog-application/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
python-jose
loguru